requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlmodel==0.0.24
chromadb==1.0.20
httpx==0.28.1
pymupdf==1.26.4
pypdf==6.0.0
beautifulsoup4==4.13.5
trafilatura==2.0.0
python-multipart==0.0.20
openai==1.102.0
python-dotenv==1.1.1
pydantic-settings==2.10.1
lxml-html-clean==0.4.2